+++ services/src/main/java/org/keycloak/protocol/oidc/AccessTokenIntrospectionProvider.java
@@ -17,32 +17,51 @@
  */
 package org.keycloak.protocol.oidc;
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.nio.charset.Charset;
+import java.util.Base64;
 import java.util.Comparator;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicReference;
 
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.apache.commons.io.IOUtils;
+import org.keycloak.broker.oidc.OIDCIdentityProvider;
+import org.keycloak.broker.oidc.OIDCIdentityProviderConfig;
+import org.keycloak.broker.provider.util.SimpleHttp;
+
 import com.fasterxml.jackson.databind.node.ObjectNode;
+import org.jboss.logging.Logger;
 import org.keycloak.TokenVerifier;
 import org.keycloak.common.VerificationException;
+import org.keycloak.common.util.Time;
+import org.keycloak.connections.httpclient.HttpClientProvider;
 import org.keycloak.crypto.SignatureProvider;
 import org.keycloak.crypto.SignatureVerifierContext;
 import org.keycloak.models.ClientModel;
+import org.keycloak.models.IdentityProviderModel;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.ProtocolMapperModel;
 import org.keycloak.models.RealmModel;
 import org.keycloak.models.UserModel;
-import org.keycloak.models.UserSessionModel;
+import org.keycloak.models.customcache.CustomCacheProvider;
+import org.keycloak.models.customcache.CustomCacheProviderFactory;
 import org.keycloak.protocol.ProtocolMapper;
 import org.keycloak.protocol.ProtocolMapperUtils;
 import org.keycloak.protocol.oidc.mappers.OIDCIntrospectionMapper;
+import org.keycloak.protocol.oidc.representations.OIDCConfigurationRepresentation;
 import org.keycloak.representations.AccessToken;
 import org.keycloak.services.Urls;
 import org.keycloak.services.managers.UserSessionCrossDCManager;
 import org.keycloak.services.util.ClientContextUtils;
 import org.keycloak.util.JsonSerialization;
+import org.keycloak.protocol.oidc.utils.Key;
 
+import javax.ws.rs.NotFoundException;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
@@ -54,15 +73,61 @@ public class AccessTokenIntrospectionProvider implements TokenIntrospectionProvi
     private final KeycloakSession session;
     private final TokenManager tokenManager;
     private final RealmModel realm;
+    private static final Logger logger = Logger.getLogger(AccessTokenIntrospectionProvider.class);
+    private static final String wellKnown = "/.well-known/openid-configuration";
+    private static final String PARAM_TOKEN = "token";
+
+    private static CustomCacheProvider tokenRelayCache;
 
     public AccessTokenIntrospectionProvider(KeycloakSession session) {
         this.session = session;
         this.realm = session.getContext().getRealm();
         this.tokenManager = new TokenManager();
+        initTokenCache();
+    }
+
+    private void initTokenCache(){
+        if(tokenRelayCache != null)
+            return;
+        CustomCacheProviderFactory factory = (CustomCacheProviderFactory)session.getKeycloakSessionFactory().getProviderFactory(CustomCacheProvider.class, "token-relay-cache");
+        if(factory == null)
+            throw new NotFoundException("Could not initate TokenRelayCacheProvider. Was not found");
+        tokenRelayCache = factory.create(session);
     }
 
     public Response introspect(String token) {
         try {
+            String[] splitToken = token.split("\\.");
+            String accessTokenStr = new String(Base64.getUrlDecoder().decode(splitToken[1]));
+            JsonNode tokenJson = new ObjectMapper().readTree(accessTokenStr);
+            String issuer = tokenJson.get("iss").asText();
+            String realmUrl = Urls.realmIssuer(session.getContext().getUri().getBaseUri(), realm.getName());
+            if (realmUrl.equals(issuer)) {
+                return introspectKeycloak(token);
+            } else {
+                if (isExpired(tokenJson.get("exp").asLong())) {
+                    ObjectNode tokenMetadata = JsonSerialization.createObjectNode();
+                    tokenMetadata.put("active", false);
+                    return Response.ok(JsonSerialization.writeValueAsBytes(tokenMetadata)).type(MediaType.APPLICATION_JSON_TYPE).build();
+                }  else {
+                    return introspectWithExternal(token, issuer, realm);
+                }
+            }
+
+        } catch (Exception e) {
+            ObjectNode tokenMetadata = JsonSerialization.createObjectNode();
+            tokenMetadata.put("active", false);
+            try {
+                return Response.ok(JsonSerialization.writeValueAsBytes(tokenMetadata)).type(MediaType.APPLICATION_JSON_TYPE).build();
+            } catch (IOException ioException) {
+                throw new RuntimeException("Error creating token introspection response.", e);
+            }
+        }
+    }
+
+    protected Response introspectKeycloak (String token) {
+        try {
+
             AccessToken accessToken = verifyAccessToken(token);
             ObjectNode tokenMetadata;
 
@@ -143,6 +208,44 @@ public class AccessTokenIntrospectionProvider implements TokenIntrospectionProvi
             user = new UserSessionCrossDCManager(session).getUserSessionWithClient(realm, token.getSessionState(), false, client.getId()).getUser();
         }
         return user;
+    }
+
+    protected Response introspectWithExternal(String token, String issuer, RealmModel realm) throws IOException {
+
+        String cachedToken = (String) tokenRelayCache.get(new Key(token, realm.getName()));
+        if(cachedToken != null)
+            return Response.ok(cachedToken).type(MediaType.APPLICATION_JSON_TYPE).build();
+
+        try {
+            IdentityProviderModel issuerIdp = realm.getIdentityProvidersStream().filter(idp -> issuer.equals(idp.getConfig().get("issuer"))).findAny().orElse(null);
+            if (issuerIdp != null) {
+                OIDCIdentityProviderConfig oidcIssuerIdp = new OIDCIdentityProviderConfig(issuerIdp);
+                OIDCIdentityProvider oidcIssuerProvider = new OIDCIdentityProvider(session, oidcIssuerIdp);
+                InputStream inputStream = session.getProvider(HttpClientProvider.class).get(new String(oidcIssuerIdp.getIssuer() + wellKnown));
+                OIDCConfigurationRepresentation rep = JsonSerialization.readValue(inputStream, OIDCConfigurationRepresentation.class);
+                if (rep.getIntrospectionEndpoint() != null) {
+                    SimpleHttp.Response response = oidcIssuerProvider.authenticateTokenRequest(SimpleHttp.doPost(rep.getIntrospectionEndpoint(), session).param(PARAM_TOKEN, token)).asResponse();
+                    if (response.getResponse().getStatusLine().getStatusCode() > 300) {
+                        ObjectNode tokenMetadata = JsonSerialization.createObjectNode();
+                        tokenMetadata.put("active", false);
+                        return Response.ok(JsonSerialization.writeValueAsBytes(tokenMetadata)).type(MediaType.APPLICATION_JSON_TYPE).build();
+                    }
+                    String responseJson = IOUtils.toString(response.getResponse().getEntity().getContent(), Charset.defaultCharset());
+                    tokenRelayCache.put(new Key(token, realm.getName()), responseJson);
+                    return Response.status(response.getResponse().getStatusLine().getStatusCode()).type(MediaType.APPLICATION_JSON_TYPE).entity(responseJson).build();
+                }
+            }
+            //if failed to find issuer in IdPs or IntrospectionEndpoint does not exist for specific Idp return false
+            ObjectNode tokenMetadata = JsonSerialization.createObjectNode();
+            tokenMetadata.put("active", false);
+            return Response.ok(JsonSerialization.writeValueAsBytes(tokenMetadata)).type(MediaType.APPLICATION_JSON_TYPE).build();
+        } catch (Exception e) {
+            throw new RuntimeException("Error creating token introspection response.", e);
+        }
+    }
+
+    private boolean isExpired(Long exp) {
+        return exp != null && exp != 0 ? Time.currentTime() > exp : false;
     }
 
     @Override