+++ CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 For Keycloak upstream changelog please see https://www.keycloak.org/docs/latest/release_notes/index.html.
 Full Keycloak upstream jira issue can be shown if filtered by Fix version. For example [Keycloak jira issue for 15.0.2 version](https://issues.redhat.com/browse/KEYCLOAK-19161?jql=project%20%3D%20keycloak%20and%20fixVersion%20%3D%2015.0.2)
 
-## [v15.0.2-rc1.0.9] - 2022-02-01
+## [v15.0.2-rc1.1.0] - 2022-02-01
 ### Added
 - Device Authorization Grant with PKCE [KEYCLOAK-9710](https://github.com/keycloak/keycloak/issues/9710)
+- external introspection endpoint [EOSC-KC-140](https://github.com/eosc-kc/keycloak/issues/140)
 ### Fixed
 - Fix scope bug in device authorization request [RCIAM-783](https://jira.argo.grnet.gr/browse/RCIAM-783)
+- Fix SAML federation enable save button bug [RCIAM-796](https://jira.argo.grnet.gr/browse/RCIAM-796)
+
 
 ## [v15.0.2-rc1.0.9] - 2022-01-28
 ### Added